+++ src/components/pages/Media/Delete/index.tsx
