+++ src/components/pages/Subjects/AllSubjects/Delete/index.tsx
