+++ src/components/pages/Subjects/Detail/index.tsx
