+++ src/components/pages/Subjects/Detail/routes/Teachers/List/Delete/index.tsx
