+++ src/components/pages/Subjects/Detail/routes/Topics/Delete/index.tsx
